requirements.txt
streamlit
pillow
numpy
matplotlib